# src/services/inventory/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /{clusterId}/user/{userId}/syncronize:
    post:
      consumes:
      - application/json
      description: Cria um ususario que esteja na base do ostern e nao exista no cluster
      parameters:
      - description: Cluster id from where retrieve users
        in: path
        name: clusterId
        required: true
        type: integer
      - description: User id registered
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: number
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Sincronize um ususario no rabbitmq
      tags:
      - User
  /{clusterId}/virtualhost:
    get:
      consumes:
      - application/json
      description: Retrieve all virtual hosts from cluster and database
      parameters:
      - description: Cluster id from where retrieve virtualhost
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retrieve all virtual hosts from cluster and database
      tags:
      - VirtualHost
    post:
      consumes:
      - application/json
      description: Import or Create a new VirtualHost
      parameters:
      - description: Cluster id from where retrieve virtualhost
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "201":
          description: Created
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Import or Create a new VirtualHost
      tags:
      - VirtualHost
...
swagger: "2.0"